requirements.txt
python-binance
typer
requests